# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Personas",
        "description": "Mostando la Lista de URI's de mi API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://PruebaTecnicaFourgen.test"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener todos los usuarios",
                "operationId": "32f813ea7a22e2b8b862ef93084426b9",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios obtenida exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "operationId": "15cb7ac620df8a65f737624b240b8976",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener el usuario autenticado",
                "operationId": "a65897868f1d4105404848067624d207",
                "responses": {
                    "200": {
                        "description": "Usuario autenticado obtenido exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar un usuario",
                "operationId": "1c2c1eb9faac63c5a992039dfc44acec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "a0ce91d753f85c9a3caee8c136d70e4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserRequest": {
                "title": "UserRequest",
                "description": "Datos necesarios para crear un usuario",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "secreto123"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}